ingestion/pyproject.toml
[project]
name = "surf-ingestion"
version = "0.1.0"
description = "Multi-Agent Orchestration Platform - Document Ingestion Pipeline"
readme = "../README.md"
license = "Apache-2.0"
authors = [{ name = "surf contributors" }]
keywords = ["ingestion", "rag", "embeddings", "azure"]
requires-python = ">=3.12"
dependencies = [
    "pymupdf",
    "azure-search-documents",
    "azure-storage-blob",
    "azure-identity",
    "openai",
    "tiktoken",
    "pydantic>=2.0",
    "click",
    "python-dotenv",
    "httpx>=0.27.0",
    "aiohttp>=3.9",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "defusedxml>=0.7",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "pyright",
    "pip-audit",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.12"
exclude = ["scripts/"]
